move trackers to its place
